testing/web-platform/tests/sanitizer-api/sanitizer.tentative.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/helper.sub.js"></script>
</head>

<body>
<script>
    test(t => {
        s = new Sanitizer();
        assert_equals("test", s.sanitizeToString("test"));
    }, "SanitizerAPI exists!");
</script>
</body>
</html>
